chore: split off mcp methods
