+++ packages/permit2-rpc-server/src/deno-server.ts
@@ -7,11 +7,9 @@ import { Permit2RpcManager } from "./core/permit2-rpc-manager.ts";
 import { RpcSelector } from "./core/rpc-selector.ts";
 import { ChainlistWsDataSource } from "./data/chainlist-ws-data-source.ts";
 import { WsLatencyTester } from "./infra/ws-latency-tester.ts";
+import { handleMcpRequest, isMcpRequest } from "./mcp/handler.ts";
 // Adjust path to point one level up from src/
 import rpcWhitelist from "../rpc-whitelist.json" with { type: "json" };
-
-// MCP SDK imports
-import { Tool } from "npm:@modelcontextprotocol/sdk@1.0.4/types.js";
 
 // Type guard to check for valid JSON-RPC request object structure
 function isValidJsonRpcRequest(obj: unknown): obj is JsonRpcRequest {
@@ -37,617 +35,6 @@ function createJsonRpcError(id: number | string | null, code: number, message: s
     id,
     error: { code, message },
   };
-}
-
-// MCP Tool Definitions for Ethereum JSON-RPC methods
-function getEthereumTools(): Tool[] {
-  return [
-    // Core Methods
-    {
-      name: "eth_getBalance",
-      description: "Returns the balance of the account at the given address",
-      inputSchema: {
-        type: "object",
-        properties: {
-          address: {
-            type: "string",
-            description: "20-byte address to check for balance",
-          },
-          blockNumber: {
-            type: "string",
-            description: "Block number or 'latest', 'earliest', 'pending'",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID (default: 1 for Ethereum mainnet)",
-          },
-        },
-        required: ["address", "blockNumber"],
-      },
-    },
-    {
-      name: "eth_getCode",
-      description: "Returns the code at a given address",
-      inputSchema: {
-        type: "object",
-        properties: {
-          address: {
-            type: "string",
-            description: "20-byte address",
-          },
-          blockNumber: {
-            type: "string",
-            description: "Block number or 'latest', 'earliest', 'pending'",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["address", "blockNumber"],
-      },
-    },
-    {
-      name: "eth_getTransactionCount",
-      description: "Returns the number of transactions sent from an address (nonce)",
-      inputSchema: {
-        type: "object",
-        properties: {
-          address: {
-            type: "string",
-            description: "20-byte address",
-          },
-          blockNumber: {
-            type: "string",
-            description: "Block number or 'latest', 'earliest', 'pending'",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["address", "blockNumber"],
-      },
-    },
-    {
-      name: "eth_getStorageAt",
-      description: "Returns the value from a storage position at a given address",
-      inputSchema: {
-        type: "object",
-        properties: {
-          address: {
-            type: "string",
-            description: "20-byte address of the storage",
-          },
-          position: {
-            type: "string",
-            description: "Position in storage",
-          },
-          blockNumber: {
-            type: "string",
-            description: "Block number or 'latest', 'earliest', 'pending'",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["address", "position", "blockNumber"],
-      },
-    },
-    {
-      name: "eth_call",
-      description: "Executes a call without creating a transaction",
-      inputSchema: {
-        type: "object",
-        properties: {
-          transaction: {
-            type: "object",
-            description: "Transaction call object",
-            properties: {
-              from: { type: "string", description: "Optional: sender address" },
-              to: { type: "string", description: "Contract address" },
-              gas: { type: "string", description: "Optional: gas limit" },
-              gasPrice: { type: "string", description: "Optional: gas price" },
-              value: { type: "string", description: "Optional: value in wei" },
-              data: { type: "string", description: "Optional: encoded data" },
-            },
-            required: ["to"],
-          },
-          blockNumber: {
-            type: "string",
-            description: "Block number or 'latest', 'earliest', 'pending'",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["transaction", "blockNumber"],
-      },
-    },
-    {
-      name: "eth_estimateGas",
-      description: "Estimates gas needed for a transaction",
-      inputSchema: {
-        type: "object",
-        properties: {
-          transaction: {
-            type: "object",
-            description: "Transaction object",
-            properties: {
-              from: { type: "string", description: "Optional: sender address" },
-              to: { type: "string", description: "Optional: recipient address" },
-              gas: { type: "string", description: "Optional: gas limit" },
-              gasPrice: { type: "string", description: "Optional: gas price" },
-              value: { type: "string", description: "Optional: value in wei" },
-              data: { type: "string", description: "Optional: encoded data" },
-            },
-          },
-          blockNumber: {
-            type: "string",
-            description: "Optional: Block number",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["transaction"],
-      },
-    },
-    {
-      name: "eth_blockNumber",
-      description: "Returns the latest block number",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-      },
-    },
-
-    // Transaction Methods
-    {
-      name: "eth_sendRawTransaction",
-      description: "Submit a signed transaction",
-      inputSchema: {
-        type: "object",
-        properties: {
-          signedTransaction: {
-            type: "string",
-            description: "Signed transaction data",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["signedTransaction"],
-      },
-    },
-    {
-      name: "eth_getTransactionByHash",
-      description: "Returns transaction details by hash",
-      inputSchema: {
-        type: "object",
-        properties: {
-          hash: {
-            type: "string",
-            description: "Transaction hash",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["hash"],
-      },
-    },
-    {
-      name: "eth_getTransactionReceipt",
-      description: "Returns the receipt of a transaction",
-      inputSchema: {
-        type: "object",
-        properties: {
-          hash: {
-            type: "string",
-            description: "Transaction hash",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["hash"],
-      },
-    },
-    {
-      name: "eth_getTransactionByBlockHashAndIndex",
-      description: "Returns transaction by block hash and index",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockHash: {
-            type: "string",
-            description: "Block hash",
-          },
-          index: {
-            type: "string",
-            description: "Transaction index in block",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockHash", "index"],
-      },
-    },
-    {
-      name: "eth_getTransactionByBlockNumberAndIndex",
-      description: "Returns transaction by block number and index",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockNumber: {
-            type: "string",
-            description: "Block number or tag",
-          },
-          index: {
-            type: "string",
-            description: "Transaction index in block",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockNumber", "index"],
-      },
-    },
-    {
-      name: "eth_getBlockTransactionCountByHash",
-      description: "Returns the number of transactions in a block by hash",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockHash: {
-            type: "string",
-            description: "Block hash",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockHash"],
-      },
-    },
-
-    // Block Methods
-    {
-      name: "eth_getBlockByHash",
-      description: "Returns block information by hash",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockHash: {
-            type: "string",
-            description: "Block hash",
-          },
-          fullTransactions: {
-            type: "boolean",
-            description: "Return full transaction objects",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockHash", "fullTransactions"],
-      },
-    },
-    {
-      name: "eth_getBlockByNumber",
-      description: "Returns block information by number",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockNumber: {
-            type: "string",
-            description: "Block number or tag",
-          },
-          fullTransactions: {
-            type: "boolean",
-            description: "Return full transaction objects",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockNumber", "fullTransactions"],
-      },
-    },
-    {
-      name: "eth_getBlockTransactionCountByNumber",
-      description: "Returns the number of transactions in a block by number",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockNumber: {
-            type: "string",
-            description: "Block number or tag",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockNumber"],
-      },
-    },
-    {
-      name: "eth_getUncleCountByBlockHash",
-      description: "Returns the number of uncles in a block by hash",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockHash: {
-            type: "string",
-            description: "Block hash",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockHash"],
-      },
-    },
-    {
-      name: "eth_getUncleCountByBlockNumber",
-      description: "Returns the number of uncles in a block by number",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockNumber: {
-            type: "string",
-            description: "Block number or tag",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockNumber"],
-      },
-    },
-    {
-      name: "eth_getUncleByBlockHashAndIndex",
-      description: "Returns uncle block by hash and index",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockHash: {
-            type: "string",
-            description: "Block hash",
-          },
-          index: {
-            type: "string",
-            description: "Uncle index",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockHash", "index"],
-      },
-    },
-    {
-      name: "eth_getUncleByBlockNumberAndIndex",
-      description: "Returns uncle block by number and index",
-      inputSchema: {
-        type: "object",
-        properties: {
-          blockNumber: {
-            type: "string",
-            description: "Block number or tag",
-          },
-          index: {
-            type: "string",
-            description: "Uncle index",
-          },
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-        required: ["blockNumber", "index"],
-      },
-    },
-
-    // Network Info Methods
-    {
-      name: "eth_protocolVersion",
-      description: "Returns the current Ethereum protocol version",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-      },
-    },
-    {
-      name: "eth_syncing",
-      description: "Returns sync status or false",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-      },
-    },
-    {
-      name: "eth_coinbase",
-      description: "Returns the coinbase address",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-      },
-    },
-    {
-      name: "eth_chainId",
-      description: "Returns the chain ID",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Override chain ID for routing",
-          },
-        },
-      },
-    },
-    {
-      name: "eth_mining",
-      description: "Returns true if client is mining",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-      },
-    },
-    {
-      name: "eth_hashrate",
-      description: "Returns the current hashrate",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-      },
-    },
-    {
-      name: "eth_gasPrice",
-      description: "Returns the current gas price",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-      },
-    },
-    {
-      name: "eth_accounts",
-      description: "Returns the list of accounts",
-      inputSchema: {
-        type: "object",
-        properties: {
-          chainId: {
-            type: "number",
-            description: "Chain ID",
-          },
-        },
-      },
-    },
-  ];
-}
-
-// Convert MCP tool arguments to JSON-RPC parameters
-function buildRpcParams(method: string, args: any): unknown[] {
-  switch (method) {
-    case "eth_getBalance":
-      return [args.address, args.blockNumber];
-    case "eth_getCode":
-      return [args.address, args.blockNumber];
-    case "eth_getTransactionCount":
-      return [args.address, args.blockNumber];
-    case "eth_getStorageAt":
-      return [args.address, args.position, args.blockNumber];
-    case "eth_call":
-      return [args.transaction, args.blockNumber];
-    case "eth_estimateGas":
-      return args.blockNumber ? [args.transaction, args.blockNumber] : [args.transaction];
-    case "eth_sendRawTransaction":
-      return [args.signedTransaction];
-    case "eth_getTransactionByHash":
-      return [args.hash];
-    case "eth_getTransactionReceipt":
-      return [args.hash];
-    case "eth_getTransactionByBlockHashAndIndex":
-      return [args.blockHash, args.index];
-    case "eth_getTransactionByBlockNumberAndIndex":
-      return [args.blockNumber, args.index];
-    case "eth_getBlockTransactionCountByHash":
-      return [args.blockHash];
-    case "eth_getBlockByHash":
-      return [args.blockHash, args.fullTransactions];
-    case "eth_getBlockByNumber":
-      return [args.blockNumber, args.fullTransactions];
-    case "eth_getBlockTransactionCountByNumber":
-      return [args.blockNumber];
-    case "eth_getUncleCountByBlockHash":
-      return [args.blockHash];
-    case "eth_getUncleCountByBlockNumber":
-      return [args.blockNumber];
-    case "eth_getUncleByBlockHashAndIndex":
-      return [args.blockHash, args.index];
-    case "eth_getUncleByBlockNumberAndIndex":
-      return [args.blockNumber, args.index];
-    case "eth_blockNumber":
-    case "eth_protocolVersion":
-    case "eth_syncing":
-    case "eth_coinbase":
-    case "eth_chainId":
-    case "eth_mining":
-    case "eth_hashrate":
-    case "eth_gasPrice":
-    case "eth_accounts":
-      return [];
-    default:
-      return [];
-  }
-}
-
-// Detect if request is MCP protocol
-function isMcpRequest(body: unknown): boolean {
-  if (typeof body === "object" && body !== null && "method" in body) {
-    const method = (body as any).method;
-    return typeof method === "string" && (
-      method === "initialize" ||
-      method.startsWith("tools/") ||
-      method.startsWith("resources/") ||
-      method.startsWith("prompts/")
-    );
-  }
-  return false;
 }
 
 const PORT = parseInt(Deno.env.get("PORT") ?? "8000");
@@ -1097,92 +484,7 @@ const handler = async (request: Request): Promise<Response> => {
 
   // Check if this is an MCP request (at root path or with chainId)
   if (isMcpRequest(requestBody)) {
-    const mcpRequest = requestBody as any;
-    let mcpResponse: any;
-
-    // Extract chainId from path if present, otherwise from request params
-    let chainId = 1; // Default to Ethereum mainnet
-    if (pathParts.length === 1) {
-      const parsed = parseInt(pathParts[0], 10);
-      if (!isNaN(parsed)) {
-        chainId = parsed;
-      }
-    }
-
-    switch (mcpRequest.method) {
-      case "initialize":
-        mcpResponse = {
-          protocolVersion: "2025-06-18",
-          capabilities: { tools: {} },
-          serverInfo: {
-            name: "ethereum-json-rpc",
-            version: "1.0.0",
-          },
-        };
-        break;
-
-      case "tools/list":
-        mcpResponse = {
-          tools: getEthereumTools(),
-        };
-        break;
-
-      case "tools/call": {
-        const toolName = mcpRequest.params?.name;
-        const toolArgs = mcpRequest.params?.arguments || {};
-
-        // Use chainId from arguments if provided, otherwise use path or default
-        if (toolArgs.chainId) {
-          chainId = toolArgs.chainId;
-        }
-
-        // Build RPC parameters
-        const params = buildRpcParams(toolName, toolArgs);
-
-        try {
-          // Execute via existing RPC manager
-          const result = await manager.send(chainId, toolName, params);
-
-          mcpResponse = {
-            content: [{
-              type: "text",
-              text: JSON.stringify(result, null, 2),
-            }],
-          };
-        } catch (e) {
-          const error = e instanceof Error ? e : new Error(String(e));
-          mcpResponse = {
-            error: {
-              code: -32603,
-              message: error.message,
-            },
-          };
-        }
-        break;
-      }
-
-      default:
-        mcpResponse = {
-          error: {
-            code: -32601,
-            message: `Method not found: ${mcpRequest.method}`,
-          },
-        };
-    }
-
-    // Return MCP response
-    return new Response(
-      JSON.stringify({
-        jsonrpc: "2.0",
-        id: mcpRequest.id,
-        result: mcpResponse.error ? undefined : mcpResponse,
-        error: mcpResponse.error,
-      }),
-      {
-        status: 200,
-        headers: { ...corsHeaders, "Content-Type": "application/json" },
-      },
-    );
+    return handleMcpRequest({ requestBody, pathParts, manager, corsHeaders });
   }
 
   // For regular JSON-RPC requests, require chainId in path
